renamed typescript files
